Removed the interface TzevaAdomNotifierConfigFactory
There is simply 1 factory class, the interface was overkill.
